perf-test-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "perf-test-agent"
version = "2.0.0"
description = "Agentic Performance Testing Framework for AT&T CTx CQE"
requires-python = ">=3.11"

dependencies = [
    # === LangChain / LLM ===
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "openai>=1.40.0",

    # === Azure ===
    "azure-identity>=1.17.0",
    "azure-search-documents>=11.5.0",
    "azure-mgmt-containerservice>=30.0.0",
    "azure-mgmt-network>=26.0.0",

    # === Data / Models ===
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "snowflake-connector-python>=3.10.0",

    # === Integrations ===
    "jira>=3.8.0",
    "httpx>=0.27.0",
    "requests>=2.32.0",
    "msal>=1.28.0",

    # === API / Web ===
    "fastapi>=0.112.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "python-multipart>=0.0.9",

    # === Utilities ===
    "python-dotenv>=1.0.0",
    "structlog>=24.2.0",
    "rich>=13.7.0",
    "tenacity>=8.4.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.4",
    "python-docx>=1.1.0",
    "croniter>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
